requirements.txt
fastapi
uvicorn[standard]
elasticsearch>=8.0.0,<9.0.0
python-dotenv
